+++ app/Http/Controllers/SubbredditsController.php
@@ -29,8 +29,8 @@ class SubbredditsController extends Controller
     {
         $subbreddit = new \App\Subbreddit;
         $subbreddit->user_id = \Auth::user()->id;
-        $subbreddit->title = $request->title;
-        $subbreddit->content = $request->content;
+        $subbreddit->name = $request->name;
+        $subbreddit->description = $request->description;
 
         $subbreddit->save();
 
@@ -61,10 +61,13 @@ class SubbredditsController extends Controller
     public function update(Request $request, $id)
     {
         $subbreddit = \App\Subbreddit::find($id);
-        $subbreddit->name = $request->name;
-        $subbreddit->description = $request->description;
-        $subbreddit->save();
-
+        if ($subbreddit->user_id == \Auth::user()->id) {
+            $subbreddit->name = $request->name;
+            $subbreddit->description = $request->description;
+            $subbreddit->save();
+        } else {
+            return response("Unauthorized", 403);
+        }
         return $subbreddit;
     }
 
@@ -77,7 +80,11 @@ class SubbredditsController extends Controller
     public function destroy($id)
     {
         $subbreddit = \App\Subbreddit::find($id);
-        $subbreddit->delete();
+        if ($subbreddit->user_id == \Auth::user()->id) {
+            $subbreddit->delete();
+        } else {
+            return response("Unauthorized", 403);
+        }
         return $subbreddit;    
     }
 }
